% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/des_fixed.R
\name{des_fixed}
\alias{des_fixed}
\title{Design a fixed-sample (single-stage) clinical trial for a normally
distributed primary outcome}
\usage{
des_fixed(
  alpha = 0.05,
  beta = 0.2,
  delta = 0.2,
  sigma0 = 1,
  sigma1 = sigma0,
  ratio = 1,
  quantile_sub = FALSE,
  integer_n = TRUE,
  summary = FALSE
)
}
\arguments{
\item{alpha}{A \code{\link{numeric}} indicating the chosen value for
\ifelse{html}{\out{<i>&alpha;</i>}}{\eqn{\alpha}}, the desired type-I
error-rate. Must be strictly between 0 and 1. Defaults to \code{0.05}.}

\item{beta}{A \code{\link{numeric}} indicating the chosen value for
\ifelse{html}{\out{<i>&beta;</i>}}{\eqn{\beta}}, the desired type-II
error-rate. Must be strictly between 0 and 1. Defaults to \code{0.2}.}

\item{delta}{A \code{\link{numeric}} indicating the chosen value for
\ifelse{html}{\out{<i>&delta;</i>}}{\eqn{\delta}}, the treatment effect to
power the trial for. Must be strictly positive. Defaults to \code{0.2}.}

\item{sigma0}{A \code{\link{numeric}} indicating the chosen value for
\ifelse{html}{\out{<i>&sigma;</i><sub>0</sub>}}{\eqn{\sigma_0}}, the
standard deviation of the responses in the control arm. Must be strictly
positive. Defaults to \code{1}.}

\item{sigma1}{A \code{\link{numeric}} indicating the chosen value for
\ifelse{html}{\out{<i>&sigma;</i><sub>1</sub>}}{\eqn{\sigma_1}}, the
standard deviation of the responses in the experimental arm. Must be strictly
positive. Defaults to \code{sigma0}.}

\item{ratio}{A \code{\link{numeric}} indicating the chosen value for
\ifelse{html}{\out{<i>r</i>}}{\eqn{r}}, the allocation ratio to the
experimental arm relative to the control arm. Must be strictly positive.
Defaults to \code{1}.}

\item{quantile_sub}{A \code{\link{logical}} variable indicating whether
quantile substitution should be applied to the identified rejection boundary.
Defaults to \code{FALSE}.}

\item{integer_n}{A \code{\link{logical}} variable indicating whether the
computed values for \ifelse{html}{\out{<i>n</i><sub>0</sub>}}{\eqn{n_0}} and
\ifelse{html}{\out{<i>n</i><sub>1</sub>}}{\eqn{n_1}}, the group sizes in the
control and experimental arms, should be forced to be whole numbers. Defaults
to \code{TRUE}.}

\item{summary}{A \code{\link{logical}} variable indicating whether a summary
of the function's progress should be printed to the console. Defaults to
\code{FALSE}.}
}
\value{
A \code{\link{list}} with additional class \code{"OptGS_des"}. It
will contain each of the input variables (subject to internal modification),
relating them to the outputs of the various group-sequential design functions
in \code{\link{OptGS}}, along with additional elements including:
\itemize{
\item \item \code{n}: A \code{\link{numeric}} giving
\ifelse{html}{\out{<i>n</i>}}{\eqn{n}}, the sample size for the identified
design.
\item \code{n0}: A \code{\link{numeric}} giving
\ifelse{html}{\out{<i>n</i><sub>0</sub>}}{\eqn{n_0}}, the sample size in the
control arm for the identified design.
\item \code{n1}: A \code{\link{numeric}} giving
\ifelse{html}{\out{<i>n</i><sub>1</sub>}}{\eqn{n_1}}, the sample size in the
experimental arm for the identified design.
\item \code{name}: A \code{\link{character}} string giving a name for the
identified design.
\item \code{opchar}: A \code{\link[tibble]{tibble}} giving the operating
characteristics of the identified design when
\ifelse{html}{\out{<i>&tau;</i> = 0}}{\eqn{\tau = 0}} and
\ifelse{html}{\out{<i>&tau;</i> = <i>&delta;</i>}}{\eqn{\tau = \delta}}.
}
}
\description{
\code{des_fixed()} determines fixed-sample (i.e., single-stage) clinical
trial designs assuming the primary outcome variable is normally distributed.
}
\examples{
# The fixed-sample design for the default parameters
des <- des_fixed()
}
\seealso{
\code{\link{des_gs}}, \code{\link{des_nearopt}},
\code{\link{des_opt}}, \code{\link{opchar}},
\code{\link{sim}}, \code{\link{plot.OptGS_des}},
\code{\link{plot.OptGS_opchar}}, \code{\link{print.OptGS_des}},
\code{\link{summary.OptGS_des}}
}
