% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/des_opt.R
\name{des_opt}
\alias{des_opt}
\title{Design an optimal group-sequential clinical trial for a normally distributed
primary outcome}
\usage{
des_opt(J = 2, alpha = 0.05, beta = 0.2, delta = 0.2, sigma0 = 1,
  sigma1 = sigma0, ratio = 1, w = c(1, 0, 0, 0), quantile_sub = F,
  integer_n = T, parallel = 1, popSize = 50, maxiter = 100,
  run = maxiter, seed = Sys.time(), summary = F)
}
\arguments{
\item{J}{A \code{\link{numeric}} indicating the chosen value for
\ifelse{html}{\out{<i>J</i>}}{\eqn{J}}, the maximal allowed number of stages.
Must be an integer greater than or equal to 2. Defaults to \code{2}.}

\item{alpha}{A \code{\link{numeric}} indicating the chosen value for
\ifelse{html}{\out{<i>&alpha;</i>}}{\eqn{\alpha}}, the desired type-I
error-rate. Must be strictly between 0 and 1. Defaults to \code{0.05}.}

\item{beta}{A \code{\link{numeric}} indicating the chosen value for
\ifelse{html}{\out{<i>&beta;</i>}}{\eqn{\beta}}, the desired type-II
error-rate. Must be strictly between 0 and 1. Defaults to \code{0.2}.}

\item{delta}{A \code{\link{numeric}} indicating the chosen value for
\ifelse{html}{\out{<i>&delta;</i>}}{\eqn{\delta}}, the treatment effect to
power the trial for. Must be strictly positive. Defaults to \code{0.2}.}

\item{sigma0}{A \code{\link{numeric}} indicating the chosen value for
\ifelse{html}{\out{<i>&sigma;</i><sub>0</sub>}}{\eqn{\sigma_0}}, the
standard deviation of the responses in the control arm. Must be strictly
positive. Defaults to \code{1}.}

\item{sigma1}{A \code{\link{numeric}} indicating the chosen value for
\ifelse{html}{\out{<i>&sigma;</i><sub>1</sub>}}{\eqn{\sigma_1}}, the
standard deviation of the responses in the experimental arm. Must be strictly
positive. Defaults to \code{sigma0}.}

\item{ratio}{A \code{\link{numeric}} indicating the chosen value for
\ifelse{html}{\out{<i>r</i>}}{\eqn{r}}, the allocation ratio to the
experimental arm relative to the control arm. Must be strictly positive.
Defaults to \code{1}.}

\item{w}{A \code{\link{numeric}} \code{\link{vector}} of length 4 indicating
the weights to use in the optimality criteria. All elements must be greater
than or equal to 0, and at least one of the first 3 elements must be strictly
positive. See the vignette for further details. Defaults to
\code{c(1, 0, 0, 0)}.}

\item{quantile_sub}{A \code{\link{logical}} variable indicating whether
quantile substitution should be applied to the identified stopping
boundaries. Defaults to \code{F}.}

\item{integer_n}{A \code{\link{logical}} variable indicating whether the
computed values for \ifelse{html}{\out{<i>n</i><sub>0</sub>}}{\eqn{n_0}} and
\ifelse{html}{\out{<i>n</i><sub>1</sub>}}{\eqn{n_1}}, the group sizes in the
control and experimental arms, should be forced to be whole numbers. Defaults
to \code{T}.}

\item{parallel}{See \code{\link[GA]{ga}}.}

\item{popSize}{See \code{\link[GA]{ga}}.}

\item{maxiter}{See \code{\link[GA]{ga}}.}

\item{run}{See \code{\link[GA]{ga}}.}

\item{seed}{A variable to be passed to \code{\link{set.seed}}, to make the
calculations reproducible.}

\item{summary}{A \code{\link{logical}} variable indicating whether a summary
of the function's progress should be printed to the console. Defaults to
\code{F}.}
}
\value{
A \code{\link{list}} with additional class \code{"OptGS_des"}. It
will contain each of the input variables (subject to internal modification),
relating to the various design functions from \code{\link[OptGS]{OptGS}},
along with the following elements:
\itemize{
\item \code{CovZ}: A \code{\link{numeric}} \code{\link{matrix}} giving
\ifelse{html}{\out{Cov(<b><i>Z</i></b>)}}{\eqn{Cov(\bold{Z})}}, the
covariance between the standardised test statistics for the identified
design.
\item \code{e}: A \code{\link{numeric}} \code{\link{vector}} giving
\ifelse{html}{\out{<b><i>e</i></b>}}{\eqn{\bold{e}}}, the efficacy stopping
boundaries for the identified design.
\item \code{f}: A \code{\link{numeric}} \code{\link{vector}} giving
\ifelse{html}{\out{<b><i>f</i></b>}}{\eqn{\bold{f}}}, the futility stopping
boundaries for the identified design.
\item \code{GA}: A \code{\link{list}} containing the output from the call to
\code{\link[GA]{ga}} and each of the \code{\link[GA]{ga}} specific input
parameters.
\item \code{I}: A \code{\link{numeric}} \code{\link{vector}} giving
\ifelse{html}{\out{<b><i>I</i></b>}}{\eqn{\bold{I}}}, the vector of
information levels for the identified design.
\item \code{n}: A \code{\link{numeric}} \code{\link{vector}} giving
\ifelse{html}{\out{<b><i>n</i></b>}}{\eqn{\bold{n}}}, the vector of
stage-wise sample sizes for the identified design.
\item \code{n0}: A \code{\link{numeric}} giving
\ifelse{html}{\out{<i>n</i><sub>0</sub>}}{\eqn{n_0}}, the group size in the
control arm for the identified design.
\item \code{n1}: A \code{\link{numeric}} giving
\ifelse{html}{\out{<i>n</i><sub>1</sub>}}{\eqn{n_1}}, the group size in the
experimental arm for the identified design.
\item \code{name}: A \code{\link{character}} string giving a name for the
identified design.
\item \code{opchar}: A \code{\link[tibble]{tibble}} giving the operating
characteristics of the identified design when
\ifelse{html}{\out{<i>&tau;</i> = 0}}{\eqn{\tau = 0}},
\ifelse{html}{\out{<i>&tau;</i> = <i>&delta;</i>}}{\eqn{\tau = \delta}}, and
\ifelse{html}{\out{<i>&tau;</i> =
argmax<sub>&theta;</sub><i>ESS</i>(<i>&theta;</i>)}}{
\eqn{\tau = argmax_{\theta}ESS(\theta)}}.
}
}
\description{
\code{des_opt()} determines optimal group-sequential clinical trial designs
assuming the primary outcome variable is normally distributed, using the
approach proposed in Wason \italic{et al} (2012).
}
\examples{
# The optimal design for the default parameters
des     <- des_opt()
# A three-stage optimal design
des_3   <- des_opt(J = 3)
# Optimal under the alternative hypothesis
des_alt <- des_opt(w = c(0, 1, 0, 0))
}
\seealso{
\code{\link{an}}, \code{\link{build}}, \code{\link{des_gs}},
\code{\link{des_nearopt}}, \code{\link{opchar}}, \code{\link{sim}},
\code{\link{plot.OptGS_des}}, \code{\link{plot.OptGS_opchar}},
\code{\link{print.OptGS_des}}, \code{\link{summary.OptGS_des}}.
}
