% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est.R
\name{est}
\alias{est}
\title{Evaluate point estimators of the treatment effect for two-stage
group-sequential clinical trial designs with a normally distributed
primary outcome}
\usage{
est(
  des = OptGS::des_gs(),
  tau = seq(-des$delta, 2 * des$delta, length.out = 100),
  density = 1000,
  summary = FALSE
)
}
\arguments{
\item{des}{An object of class OptGS_des, as returned by \code{\link{des_gs}},
\code{\link{des_nearopt}}, \code{\link{des_opt}}, or \code{\link{build}}.
Defaults to \code{des_gs()}. Currently, this must be a two-stage trial.}

\item{tau}{A \code{\link{numeric}} \code{\link{vector}} indicating the values
of \ifelse{html}{\out{<i>&tau;</i>}}{\eqn{\tau}} to perform calculations for.
Defaults to \code{seq(-des$delta, 2*des$delta, length.out = 100)}.}

\item{density}{A \code{\link{numeric}} whole number indicating the number of
standardised test statistics to consider when computing the bias and RMSE.
Larger values increase accuracy of integral evaluations but also increase run
time. Defaults to \code{1000}.}

\item{summary}{A \code{\link{logical}} variable indicating whether a summary
of the function's progress should be printed to the console. Defaults to
\code{FALSE}.}
}
\value{
A \code{\link{list}} with additional class \code{"OptGS_est"}. It
will contain each of the input variables (subject to internal modification),
along with the following elements:
\itemize{
\item \code{estimates}: A \code{\link{tibble}} giving details on point
estimates.
\item \code{estimators}: A \code{\link{tibble}} giving performance of the
point estimators.
}
}
\description{
\code{est()} computes unadjusted and adjusted point estimates of the
treatment effect based on results from a two-stage group-sequential clinical
trial. It uses these to determine point estimator (conditional and marginal)
bias and residual mean squared error.
}
\examples{
# Calculations for the default parameters
perf <- est()
}
\seealso{
\code{\link{build}}, \code{\link{des_gs}},
\code{\link{des_nearopt}}, \code{\link{des_opt}}, \code{\link{sim}}.
}
