% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui.R
\name{gui}
\alias{gui}
\title{Graphical user interface to optimal group sequential trial design
determination}
\usage{
gui()
}
\description{
\code{gui()} run an R Shiny web browser based graphical user interface for
running the functions \code{\link{des_gs}}, \code{\link{des_nearopt}},
\code{\link{opchar}}, and \code{\link{plot.des_OptGS}}.
}
\examples{
# Launch the GUI
\dontrun{
gui()
}
}
\seealso{
\code{\link{des_gs}}, \code{\link{des_nearopt}},
\code{\link{opchar}}, \code{\link{plot.des_OptGS}}.
}
