% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.OptGS_des.R
\name{plot.OptGS_des}
\alias{plot.OptGS_des}
\title{Plot the stopping boundaries of a group-sequential clinical trial design for
a normally distributed primary outcome}
\usage{
\method{plot}{OptGS_des}(x, ..., output = F)
}
\arguments{
\item{x}{An object of class OptGS_des, as returned by \code{\link{des_gs}},
\code{\link{des_nearopt}}, \code{\link{des_opt}}, or \code{\link{build}}.}

\item{...}{For compatability with the generic.}

\item{output}{A \code{\link{logical}} variable indicating whether to return
output from the function.}
}
\value{
A \code{\link{list}} containing each of the input variables (subject
to internal modification), along with the following elements:
\itemize{
\item \code{plots}: A \code{\link{list}} containing the stopping boundary
plotting objects.
}
}
\description{
\code{plot.OptGS_des()} plots the stopping boundaries from a group-sequential
clinical trial design object for a normally distributed primary outcome.
Plots are produced by stage and by sample size.
}
\examples{
# A two-stage design's boundaries
plot(opt_gs())
# A three-stage design's boundaries
plot(opt_gs(J = 3))
}
\seealso{
\code{\link{an}}, \code{\link{des_gs}}, \code{\link{des_nearopt}},
\code{\link{des_opt}}, \code{\link{build}}, \code{\link{sim}},
\code{\link{opchar}}, \code{\link{plot.OptGS_opchar}},
\code{\link{print.OptGS_des}}, \code{\link{summary.OptGS_des}}.
}
