% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.OptGS_est.R
\name{plot.OptGS_est}
\alias{plot.OptGS_est}
\title{Plot point estimator performance for two-stage group-sequential clinical
trial designs with a normally distributed primary outcome}
\usage{
\method{plot}{OptGS_est}(x, ..., output = FALSE)
}
\arguments{
\item{x}{An object of class OptGS_est, as returned by \code{\link{est}}.}

\item{...}{For compatability with the generic.}

\item{output}{A \code{\link{logical}} variable indicating whether to return
output from the function.}
}
\value{
A \code{\link{list}} containing each of the input variables (subject
to internal modification), along with the following elements:
\itemize{
\item \code{plots}: A \code{\link{list}} containing the available plots.
}
}
\description{
\code{plot.OptGS_est()} plots (conditional and marginal) point estimator bias
and residual mean squared error for a specified two-stage group-sequential
clinical trial design.
}
\examples{
# For the default two-stage design
plot(est(des = opt_gs(), tau = seq(-des$delta, 2*des$delta,
         length.out = 100)))
}
\seealso{
\code{\link{est}}
}
