% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.OptGS_opchar.R
\name{plot.OptGS_opchar}
\alias{plot.OptGS_opchar}
\title{Plot the operating characteristics of a group-sequential clinical trial
design for a normally distributed primary outcome}
\usage{
\method{plot}{OptGS_opchar}(x, ..., output = F)
}
\arguments{
\item{x}{An object of class OptGS_opchar, as returned by
\code{\link{opchar}}.}

\item{...}{For compatability with the generic.}

\item{output}{A \code{\link{logical}} variable indicating whether to return
output from the function.}
}
\value{
A \code{\link{list}} containing each of the input variables (subject
to internal modification), along with the following elements:
\itemize{
\item \code{plots}: A \code{\link{list}} containing the operating
characteristic plotting objects.
}
}
\description{
\code{plot.OptGS_opchar()} plots values from a group-sequential clinical
trial operating characteristic object for a normally distributed primary
outcome.
}
\examples{
# A two-stage design's operating characteristics
plot(opchar(opt_gs(), tau = seq(-des$delta, 2*des$delta, length.out = 100)))
# A three-stage design's operating characteristics
plot(opchar(opt_gs(J = 3),
     tau = seq(-des$delta, 2*des$delta, length.out = 100)))
}
\seealso{
\code{\link{an}}, \code{\link{des_gs}}, \code{\link{des_nearopt}},
\code{\link{des_opt}}, \code{\link{build}}, \code{\link{sim}},
\code{\link{opchar}}, \code{\link{plot.OptGS_des}},
\code{\link{print.OptGS_des}}, \code{\link{summary.OptGS_des}}.
}
