% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.OptGS_des.R
\name{summary.OptGS_des}
\alias{summary.OptGS_des}
\title{Summarise the key details of a group-sequential clinical trial design for a
normally distributed primary outcome}
\usage{
\method{summary}{OptGS_des}(object, ...)
}
\arguments{
\item{object}{An object of class OptGS_des, as returned by
\code{\link{des_gs}}, \code{\link{des_nearopt}}, \code{\link{des_opt}}, or
\code{\link{build}}. Defaults to \code{des_gs()}.}

\item{...}{For compatability with the generic.}
}
\description{
\code{summary.OptGS_des()} summarises the key details from a group-sequential
clinical trial design object for a normally distributed primary outcome.
}
\examples{
# A two-stage design
summary(opt_gs())
# A three-stage design
summary(opt_gs(J = 3))
}
\seealso{
\code{\link{an}}, \code{\link{des_gs}}, \code{\link{des_nearopt}},
\code{\link{des_opt}}, \code{\link{build}}, \code{\link{sim}},
\code{\link{plot.OptGS_des}}, \code{\link{plot.OptGS_opchar}},
\code{\link{opchar}}, \code{\link{summary.OptGS_des}}.
}
