% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim}
\alias{sim}
\title{Simulate group-sequential clinical trials for a normally distributed primary
outcome}
\usage{
sim(
  des = des_gs(),
  tau = seq(-des$delta, 2 * des$delta, length.out = 100),
  estimators = FALSE,
  replicates = 1000,
  summary = FALSE
)
}
\arguments{
\item{des}{An object of class \code{"OptGS_des"}, as returned by
\code{\link{des_gs}}, \code{\link{des_nearopt}}, \code{\link{des_opt}}, or
\code{\link{build}}. Defaults to \code{des_gs()}.}

\item{tau}{A \code{\link{numeric}} \code{\link{vector}} indicating the values
of \ifelse{html}{\out{<i>&tau;</i>}}{\eqn{\tau}} to perform simulations for.
Defaults to \code{des$opchar$tau}.}

\item{replicates}{A \code{\link{numeric}} indicating the number of replicate
simulations to use for each value of
\ifelse{html}{\out{<i>&tau;</i>}}{\eqn{\tau}}. Must be a whole number.
Defaults to \code{1000}.}

\item{summary}{A \code{\link{logical}} variable indicating whether a summary
of the function's progress should be printed to the console. Defaults to
\code{FALSE}.}
}
\value{
A \code{\link{list}} with additional class \code{"OptGS_sim"}. It
will contain each of the input variables (subject to internal modification),
along with the following element:
\itemize{
\item \code{sim}: A \code{\link{tibble}} giving the estimated operating
characteristics.
}
}
\description{
\code{sim()} simulates group-sequential clinical trials in order to validate
certain analytically determined operating characteristics (as returned by
\code{\link{opchar}}) .
}
\examples{
# Simulations for the default parameters
sim     <- sim()
}
\seealso{
\code{\link{build}}, \code{\link{des_gs}},
\code{\link{des_nearopt}}, \code{\link{des_opt}}, \code{\link{opchar}}
}
